GH-2033: Move StoreConnection to oaj.tdb1.sys
