+++ jena-tdb1/src/test/java/org/apache/jena/tdb1/setup/TestStoreParamsCreate.java
@@ -27,8 +27,8 @@ import org.apache.jena.atlas.json.JSON ;
 import org.apache.jena.atlas.json.JsonObject ;
 import org.apache.jena.atlas.lib.FileOps ;
 import org.apache.jena.tdb1.ConfigTest;
-import org.apache.jena.tdb1.StoreConnection;
 import org.apache.jena.tdb1.base.file.Location;
+import org.apache.jena.tdb1.sys.StoreConnection;
 import org.junit.After ;
 import org.junit.Before ;
 import org.junit.Test ;
