+++ jena-tdb1/src/test/java/org/apache/jena/tdb1/transaction/AbstractTestTransSeq.java
@@ -28,9 +28,9 @@ import org.apache.jena.sparql.JenaTransactionException;
 import org.apache.jena.sparql.core.DatasetGraph ;
 import org.apache.jena.sparql.core.Quad ;
 import org.apache.jena.sparql.sse.SSE ;
-import org.apache.jena.tdb1.StoreConnection;
 import org.apache.jena.tdb1.TDB1Exception;
 import org.apache.jena.tdb1.store.DatasetGraphTDB;
+import org.apache.jena.tdb1.sys.StoreConnection;
 import org.apache.jena.tdb1.sys.SystemTDB;
 import org.junit.AfterClass ;
 import org.junit.BeforeClass ;
