+++ jena-tdb1/src/test/java/org/apache/jena/tdb1/transaction/TestTransRestart.java
@@ -30,7 +30,6 @@ import org.apache.jena.query.TxnType;
 import org.apache.jena.sparql.core.Quad ;
 import org.apache.jena.sparql.sse.SSE ;
 import org.apache.jena.tdb1.ConfigTest;
-import org.apache.jena.tdb1.StoreConnection;
 import org.apache.jena.tdb1.base.block.FileMode;
 import org.apache.jena.tdb1.base.file.FileFactory;
 import org.apache.jena.tdb1.base.file.Location;
@@ -38,6 +37,7 @@ import org.apache.jena.tdb1.base.objectfile.ObjectFile;
 import org.apache.jena.tdb1.setup.DatasetBuilderStd;
 import org.apache.jena.tdb1.store.DatasetGraphTDB;
 import org.apache.jena.tdb1.sys.Names;
+import org.apache.jena.tdb1.sys.StoreConnection;
 import org.apache.jena.tdb1.sys.SystemTDB;
 import org.apache.jena.tdb1.sys.TDBInternal;
 import org.junit.After ;
